source/requirements.txt
jinja2==2.8
PyYAML==3.12